requirements.txt
# requirements.txt

Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
python-dotenv==1.0.1
google-generativeai==0.7.1
google-cloud-texttospeech==2.16.2
google-cloud-storage==2.17.0
